Reorganize gateway chain API code so it is reusable, make ask about Infura when calling gateway create
