+++ hummingbot/client/command/gateway_api_manager.py
@@ -1,0 +1,99 @@
+import json
+from contextlib import contextmanager
+from enum import Enum
+from typing import TYPE_CHECKING, Generator
+
+import aiohttp
+
+if TYPE_CHECKING:
+    from hummingbot.client.hummingbot_application import HummingbotApplication
+
+
+class Chain(Enum):
+    ETHEREUM = 0
+    AVALANCHE = 1
+
+
+@contextmanager
+def begin_placeholder_mode(hb: "HummingbotApplication") -> Generator["HummingbotApplication", None, None]:
+    hb.app.clear_input()
+    hb.placeholder_mode = True
+    hb.app.hide_input = True
+    try:
+        yield hb
+    finally:
+        hb.app.to_stop_config = False
+        hb.placeholder_mode = False
+        hb.app.hide_input = False
+        hb.app.change_prompt(prompt=">>> ")
+
+
+class GatewayChainApiManager:
+    """
+    Manage and test connections from gateway to chain APIs like Infura and
+    Moralis.
+    """
+
+    async def _test_evm_node(self, url_with_api_key: str) -> bool:
+        """
+        Verify that the Infura API Key is valid. If it is an empty string,
+        ignore it, but let the user know they cannot connect to ethereum.
+        """
+        async with aiohttp.ClientSession() as tmp_client:
+            headers = {"Content-Type": "application/json"}
+            data = {
+                "jsonrpc": "2.0",
+                "id": 1,
+                "method": "eth_blockNumber",
+                "params": []
+            }
+
+            resp = await tmp_client.post(url=url_with_api_key,
+                                         data=json.dumps(data),
+                                         headers=headers)
+            success = resp.status == 200
+            if success:
+                self.notify("The API Key works.")
+            else:
+                self.notify("Error occured verifying the API Key. Please check your API Key and try again.")
+            return success
+
+    async def _get_api_key(self, chain: Chain) -> str:
+        """
+        Generalize getting and testing the API key from user input
+        """
+        with begin_placeholder_mode(self):
+            while True:
+                if chain == Chain.ETHEREUM:
+                    service = 'Infura'
+                    chain_name = 'Ethereum'
+                    service_url = 'infura.io'
+                elif chain == Chain.AVALANCHE:
+                    service = 'Moralis'
+                    chain_name = 'Avalanche'
+                    service_url = 'moralis.io'
+
+                api_key: str = await self.app.prompt(prompt=f"Enter {service} API Key (required for {chain_name} node, "
+                                                            f"if you do not have one, make an account at {service_url})"
+                                                            f", otherwise configure gateway after creation:  >>> ")
+
+                self.app.clear_input()
+
+                if self.app.to_stop_config:
+                    self.app.to_stop_config = False
+                    return ''
+                try:
+                    api_key = api_key.strip()  # help check for an empty string which is valid input
+                    if api_key is None or api_key == "":
+                        self.notify(f"Setting up gateway without an {chain_name} node.")
+                    else:
+                        if chain == Chain.ETHEREUM:
+                            api_url = f"https://mainnet.infura.io/v3/{api_key}"
+                        elif chain == Chain.AVALANCHE:
+                            api_url = f"https://speedy-nodes-nyc.moralis.io/{api_key}/avalanche/mainnet"
+                        # if this throws an error, it will give the user a chance to enter another key
+                        await self._test_evm_node(api_url)
+                    return api_key
+                except Exception:
+                    self.notify(f"Error occur calling the API route: {api_url}.")
+                    raise
